study/requirements.txt
flask >= 2.2.2, < 2.3
flask-s3 >=0.3.3, < 0.4
gunicorn >= 20.1.0, < 20.2
python-dotenv >= 0.21.0, < 0.22
whitenoise >= 6.2.0 , < 6.3
